pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "verse-slides"
version = "0.3.0"
description = "CLI tool to fetch scripture and generate presentation-ready PDFs"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Chris Shinkle", email = "chris@example.com"}
]
dependencies = [
    "requests>=2.28.0",
    "reportlab>=4.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "responses>=0.22.0",
]

[tool.setuptools]
packages = ["verse_slides"]

[project.scripts]
verse-slides = "verse_slides.cli:main"
vs = "verse_slides.cli:main"
